third_party/WebKit/LayoutTests/animations/animations-csstext.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function(){
    var div = document.createElement('div');
    div.style.animationName = 'anim';
    assert_equals(div.style.cssText, 'animation-name: anim;');
    div.style.animation = 'hello 2s';
    assert_equals(div.style.cssText, 'animation: hello 2s;');
}, 'cssText serializes animation properties sanely');
</script>
